Move virtual machine config to configuration mod
It is much more proper space for the configuration class.
