+++ scripts/launcher/lib/virtual_controller.py
@@ -40,221 +40,19 @@ from pylorax import setup_logging
 from pylorax.monitor import LogMonitor
 from pylorax.mount import IsoMountpoint
 
+from .configuration import VirtualConfiguration
 from .validator import replace_new_lines
 
 import logging
 log = logging.getLogger("livemedia-creator")
 
 
-__all__ = ["VirtualConfiguration", "VirtualManager", "InstallError"]
+__all__ = ["VirtualManager", "InstallError"]
 
 
 class InstallError(Exception):
     def __str__(self):
         return super().__str__().replace("#012", "\n")
-
-
-class VirtualConfiguration(object):
-
-    def __init__(self, iso_path, ks_paths):
-        super().__init__()
-
-        self._test_name = ""
-        self._iso = iso_path
-        self._ks_paths = ks_paths
-        self._disk = []
-        self._nic = []
-        self._proxy = None
-        self._location = None
-        self._boot = None
-        self._logfile = "./livemedia.log"
-        self._tmp = "/var/tmp"
-        self._keep_image = True
-        self._ram = 1024
-        self._vnc = None
-        self._kernel_args = None
-        self._timeout = None
-
-    def __str__(self):
-        msg = "Virtual Configuration:"
-
-        for p in dir(VirtualConfiguration):
-            if p.startswith("_"):
-                continue
-            else:
-                if getattr(self, p):
-                    msg += " {}: {},".format(p, getattr(self, p))
-
-        msg = msg[:-1]
-
-        return msg
-
-    @property
-    def test_name(self):
-        """Name of this test run.
-
-        Optional. Can be empty string.
-        """
-        return self._test_name
-
-    @test_name.setter
-    def test_name(self, value):
-        """Set name of this test run.
-
-        Optional. Can be empty string.
-        """
-        self._test_name = value
-
-    @property
-    def iso_path(self) -> str:
-        """Anaconda installation .iso path to use for virt-install"""
-        return self._iso
-
-    @iso_path.setter
-    def iso_path(self, value: str):
-        """Set Anaconda installation .iso path to use for virt-install"""
-        self._iso = value
-
-    @property
-    def ks_paths(self) -> []:
-        """Kickstart file defining the install"""
-        return self._ks_paths
-
-    @ks_paths.setter
-    def ks_paths(self, value: []):
-        """Set kickstart file defining the install"""
-        self._ks_paths = value
-
-    @property
-    def disk_paths(self) -> []:
-        """Pre-existing disk image to use for destination"""
-        return self._disk
-
-    @disk_paths.setter
-    def disk_paths(self, value: []):
-        """Set pre-existing disk image to use for destination"""
-        self._disk = value
-
-    @property
-    def networks(self) -> []:
-        """Network devices to be used"""
-        return self._nic
-
-    @networks.setter
-    def networks(self, value: []):
-        """Set network devices to be used"""
-        self._nic = value
-
-    @property
-    def proxy(self) -> str:
-        """Proxy URL to use for the install"""
-        return self._proxy
-
-    @proxy.setter
-    def proxy(self, value: str):
-        """Set proxy URL to use for the install"""
-        self._proxy = value
-
-    @property
-    def location(self) -> str:
-        """Location of iso directory tree"""
-        return self._location
-
-    @location.setter
-    def location(self, value: str):
-        """Set location of iso directory tree"""
-        self._location = value
-
-    @property
-    def boot_image(self) -> str:
-        """Alternative boot image
-
-         eg. ipxe.lkrn for ibft
-         """
-        return self._boot
-
-    @boot_image.setter
-    def boot_image(self, value: str):
-        """Set alternative boot image"""
-        self._boot = value
-
-    @property
-    def log_path(self) -> str:
-        """Path to logfile"""
-        return self._logfile
-
-    @log_path.setter
-    def log_path(self, value: str):
-        """Set path to logfile"""
-        self._logfile = value
-
-    @property
-    def install_logpath(self):
-        """Virtual log file
-
-        Will be dynamically created based on the temp position.
-        """
-        return os.path.join(self._tmp, "virt-install.log")
-
-    @property
-    def temp_dir(self) -> str:
-        """Top level temporary directory"""
-        return self._tmp
-
-    @temp_dir.setter
-    def temp_dir(self, value: str):
-        """Set top level temporary directory"""
-        self._tmp = value
-
-    @property
-    def keep_image(self) -> bool:
-        """Keep raw disk image after .iso creation"""
-        return self._keep_image
-
-    @keep_image.setter
-    def keep_image(self, value: bool):
-        """Set keep raw disk image after .iso creation"""
-        self._keep_image = value
-
-    @property
-    def ram(self) -> int:
-        """Memory to allocate for installer in megabytes"""
-        return self._ram
-
-    @ram.setter
-    def ram(self, value: int):
-        """Set memory to allocate for installer in megabytes"""
-        self._ram = value
-
-    @property
-    def vnc(self) -> str:
-        """Passed to --graphics command"""
-        return self._vnc
-
-    @vnc.setter
-    def vnc(self, value: str):
-        """Set passed to --graphics command"""
-        self._vnc = value
-
-    @property
-    def kernel_args(self) -> str:
-        """Additional argument to pass to the installation kernel"""
-        return self._kernel_args
-
-    @kernel_args.setter
-    def kernel_args(self, value: str):
-        """Set additional argument to pass to the installation kernel"""
-        self._kernel_args = value
-
-    @property
-    def timeout(self) -> int:
-        """Cancel installer after X minutes"""
-        return self._timeout
-
-    @timeout.setter
-    def timeout(self, value: int):
-        """Set cancel installer after X minutes"""
-        self._timeout = value
 
 
 class VirtualInstall(object):
